# NodeJsLibraryApi/models/book.js
class Book {
  constructor({ id, title, author, isbn }) {
    this.id = id;
    this.title = title;
    this.author = author;
    this.isbn = isbn;
    this.isReserved = false;
    this.reservedBy = null;
  }
}

module.exports = Book;